refactor(gateway): Move RequiredRoles to annotation package
Contributes to 'profile' functionality
