+++ mp-gateway/src/main/java/ru/pavbatol/myplace/gateway/profile/geo/country/controller/AdminCountryController.java
@@ -10,7 +10,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import ru.pavbatol.myplace.gateway.app.api.ApiResponse;
-import ru.pavbatol.myplace.gateway.app.access.RequiredRoles;
+import ru.pavbatol.myplace.gateway.app.annotation.RequiredRoles;
 import ru.pavbatol.myplace.gateway.profile.geo.country.service.CountryService;
 import ru.pavbatol.myplace.shared.dto.profile.geo.country.CountryDto;
 import ru.pavbatol.myplace.shared.util.Marker;
