+++ mp-gateway/src/main/java/ru/pavbatol/myplace/gateway/profile/geo/house/controller/AdminHoseController.java
@@ -8,7 +8,7 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import ru.pavbatol.myplace.gateway.app.access.RequiredRoles;
+import ru.pavbatol.myplace.gateway.app.annotation.RequiredRoles;
 import ru.pavbatol.myplace.gateway.app.api.ApiResponse;
 import ru.pavbatol.myplace.gateway.profile.geo.house.service.HouseService;
 import ru.pavbatol.myplace.shared.dto.profile.geo.house.HouseDto;
